Let's start with content and no CSS
It adds all the text and markup, but there is no CSS. Which,
incidentally I don't think it is a bad idea. Should not happen but if
network errors make the CSS don't load, at least the site will look
nice, with some 90s college teacher vibes. Now I want to see this page
render on a Windows 98 virtual machine.